moved the util function
